refactor: switch from function based to class based design  and rename kso2 to kso
