+++ src/kso/Inference_latency.py
@@ -7,157 +7,158 @@ import platform
 import subprocess
 import logging
 from torchinfo import summary
-from .trainer import internal_model
+from .trainer import ExperimentManager
 from mlflow.pyfunc import PyFuncModel
 import psutil
 from pathlib import Path
 
 
-def model_stats(model, image_path, batch_size):
-    if not isinstance(model, PyFuncModel):
-        raise TypeError(f"model {model} is not PyFuncModel model")
-    device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-    torch_model = internal_model(model).to(device).eval()
-    if not isinstance(torch_model, torch.nn.Module):
-        raise ValueError(f"model {torch_model} is not a nn.Module")
-    torch_model.requires_grad_(False)
-
-    x = torchvision.io.read_image(path=image_path)
-    x = x.float() / 255.0
-    print(f"batch_size: {batch_size}")
-    x = x.unsqueeze(0).expand(batch_size, -1, -1, -1)  # (batch_size, 3, H, W)
-    x = torch.nn.functional.interpolate(
-        x, size=(640, 640), mode="bilinear", align_corners=False
-    )
-    x = x.to(device, non_blocking=False)
-
-    macs, params = profile(torch_model, inputs=(x,), verbose=False)
-    print(f"model macs: {macs:.2e}")
-    print(f"model flops: {macs * 2:.2e}")
-    print(f"model params: {params:.2e}")
-    return macs, params
+train = ExperimentManager()
 
 
-def image_num_indir(path: str):
-    VALID_EXTS = {".jpg", ".jpeg", ".png", ".tif", ".tiff", ".bmp", ".webp"}
-    # Travers all the branch of a specified path
-    files = os.listdir(path)
-    return sum([file.endswith(s) for s in VALID_EXTS for file in files])
+class ModelProfiler:
+    def __init__(self):
+        pass
 
+    def model_stats(self, model, image_path, batch_size):
+        if not isinstance(model, PyFuncModel):
+            raise TypeError(f"model {model} is not PyFuncModel model")
+        device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
+        torch_model = train.internal_model(model).to(device).eval()
+        if not isinstance(torch_model, torch.nn.Module):
+            raise ValueError(f"model {torch_model} is not a nn.Module")
+        torch_model.requires_grad_(False)
 
-def os_system():
-    "retrieve the information about the platform on which the program is running"
-    paltform = platform.uname()
-    system = paltform.system
-    return system
-
-
-def hardware_flops():
-
-    TFLOPS = 0.0
-    system = os_system()
-    if system == "Linux" and torch.cuda.is_available():
-        cu_output = subprocess.check_output(
-            "rocminfo | grep -i 'Compute Unit' | awk '{s+=$3} END{print s}'",
-            shell=True,
-            text=True,
+        x = torchvision.io.read_image(path=image_path)
+        x = x.float() / 255.0
+        print(f"batch_size: {batch_size}")
+        x = x.unsqueeze(0).expand(batch_size, -1, -1, -1)  # (batch_size, 3, H, W)
+        x = torch.nn.functional.interpolate(
+            x, size=(640, 640), mode="bilinear", align_corners=False
         )
-        CUs = int(cu_output.strip())
+        x = x.to(device, non_blocking=False)
 
-        # MI250X official specs
-        FP32_PER_GCD_TFLOPS = 47.9
-        CUS_PER_GCD = 110
+        macs, params = profile(torch_model, inputs=(x,), verbose=False)
+        print(f"model macs: {macs:.2e}")
+        print(f"model flops: {macs * 2:.2e}")
+        print(f"model params: {params:.2e}")
+        return macs, params
 
-        FP32_PER_CU_TFLOPS = FP32_PER_GCD_TFLOPS / CUS_PER_GCD
-        TFLOPS = CUs * FP32_PER_CU_TFLOPS
-    elif system == "Darwin":
-        # Apple M1/M2 GPU
-        try:
-            cores_output = subprocess.check_output(
-                "system_profiler SPDisplaysDataType | grep 'Total Number of Cores' | awk '{print $NF}'",
+    def image_num_indir(self, path: str):
+        VALID_EXTS = {".jpg", ".jpeg", ".png", ".tif", ".tiff", ".bmp", ".webp"}
+        # Travers all the branch of a specified path
+        files = os.listdir(path)
+        return sum([file.endswith(s) for s in VALID_EXTS for file in files])
+
+    def os_system(self):
+        "retrieve the information about the platform on which the program is running"
+        paltform = platform.uname()
+        system = paltform.system
+        return system
+
+    def hardware_flops(self):
+
+        TFLOPS = 0.0
+        system = self.os_system()
+        if system == "Linux" and torch.cuda.is_available():
+            cu_output = subprocess.check_output(
+                "rocminfo | grep -i 'Compute Unit' | awk '{s+=$3} END{print s}'",
                 shell=True,
                 text=True,
             )
-            cores = int(cores_output.strip())
-        except Exception as e:
-            logging.warning(f"Could not detect GPU cores: {e}")
-            cores = 0
+            CUs = int(cu_output.strip())
 
-        TFLOPS_PER_CORE = 0.325  # M1 GPU approximation
-        TFLOPS = cores * TFLOPS_PER_CORE
-    # elif system=="Windows":
-    #     try:
-    #         cores_output = subprocess.check_output(
-    #         "powershell -Command \"(Get-CimInstance Win32_Processor).NumberOfCores\"",
-    #         shell=True,
-    #         text=True
-    #         )
-    #         cores = int(cores_output.strip())
-    #     except Exception as e:
-    #         logging.warning(f"Could not detect GPU cores: {e}")
-    # logging.info(f"Detected Compute Units: {CUs}")
-    # logging.info(f"Estimated FP32 peak TFLOPS of your Slurm allocation: {TFLOPS:.2f}")
-    return TFLOPS
+            # MI250X official specs
+            FP32_PER_GCD_TFLOPS = 47.9
+            CUS_PER_GCD = 110
 
+            FP32_PER_CU_TFLOPS = FP32_PER_GCD_TFLOPS / CUS_PER_GCD
+            TFLOPS = CUs * FP32_PER_CU_TFLOPS
+        elif system == "Darwin":
+            # Apple M1/M2 GPU
+            try:
+                cores_output = subprocess.check_output(
+                    "system_profiler SPDisplaysDataType | grep 'Total Number of Cores' | awk '{print $NF}'",
+                    shell=True,
+                    text=True,
+                )
+                cores = int(cores_output.strip())
+            except Exception as e:
+                logging.warning(f"Could not detect GPU cores: {e}")
+                cores = 0
 
-def model_latency_inference(model, image_path, batch_size):
+            TFLOPS_PER_CORE = 0.325  # M1 GPU approximation
+            TFLOPS = cores * TFLOPS_PER_CORE
+        # elif system=="Windows":
+        #     try:
+        #         cores_output = subprocess.check_output(
+        #         "powershell -Command \"(Get-CimInstance Win32_Processor).NumberOfCores\"",
+        #         shell=True,
+        #         text=True
+        #         )
+        #         cores = int(cores_output.strip())
+        #     except Exception as e:
+        #         logging.warning(f"Could not detect GPU cores: {e}")
+        # logging.info(f"Detected Compute Units: {CUs}")
+        # logging.info(f"Estimated FP32 peak TFLOPS of your Slurm allocation: {TFLOPS:.2f}")
+        return TFLOPS
 
-    macs, _ = model_stats(model, image_path, batch_size)  # model FLOPs
-    FLOPs = macs * 2
+    def model_latency_inference(self, model, image_path, batch_size):
 
-    device_TFLOPs = hardware_flops()  # MI250X allocation
-    device_FLOPs = device_TFLOPs * 1e12
+        macs, _ = self.model_stats(model, image_path, batch_size)  # model FLOPs
+        FLOPs = macs * 2
 
-    latency_sec = FLOPs / device_FLOPs
-    latency_ms = latency_sec * 1e3
+        device_TFLOPs = self.hardware_flops()  # MI250X allocation
+        device_FLOPs = device_TFLOPs * 1e12
 
-    # print(f"Theoretical lower-bound latency: {latency_ms:.4f} ms")
-    return latency_ms
+        latency_sec = FLOPs / device_FLOPs
+        latency_ms = latency_sec * 1e3
 
+        # print(f"Theoretical lower-bound latency: {latency_ms:.4f} ms")
+        return latency_ms
 
-def inference_memory(model, batch_size):
-    """estimate the model memory used during inference"""
-    my_device = "cuda" if torch.cuda.is_available() else "cpu"
-    pytorch_model = internal_model(model)
-    pytorch_model.eval().to(my_device)
+    def inference_memory(self, model, batch_size):
+        """estimate the model memory used during inference"""
+        my_device = "cuda" if torch.cuda.is_available() else "cpu"
+        pytorch_model = train.internal_model(model)
+        pytorch_model.eval().to(my_device)
 
-    stats = summary(
-        pytorch_model,
-        input_size=(batch_size, 3, 640, 640),
-        col_names=("input_size", "output_size", "num_params", "mult_adds"),
-        verbose=0,
-        mode="eval",
-    )
+        stats = summary(
+            pytorch_model,
+            input_size=(batch_size, 3, 640, 640),
+            col_names=("input_size", "output_size", "num_params", "mult_adds"),
+            verbose=0,
+            mode="eval",
+        )
 
-    MB = 1e6
+        MB = 1e6
 
-    inference_bytes = (
-        stats.total_input  # input bytes
-        + stats.total_output_bytes / 2  # forward activations only
-        + stats.total_param_bytes  # model weights
-    )
+        inference_bytes = (
+            stats.total_input  # input bytes
+            + stats.total_output_bytes / 2  # forward activations only
+            + stats.total_param_bytes  # model weights
+        )
 
-    inference_mb = inference_bytes / MB
-    # print(f"Inference Memory (MB): {inference_mb:.2f}")
-    return inference_mb
+        inference_mb = inference_bytes / MB
+        # print(f"Inference Memory (MB): {inference_mb:.2f}")
+        return inference_mb
 
+    def memory_estimator(self, model, image_path, batch_size=16):
+        """estimate the inference latency and the memory usage
+        comenpared to the current slurm allocation"""
+        image_path = Path(image_path).expanduser()
+        if "SLURM_JOB_ID" in os.environ:
+            ram_gb = int(os.environ["SLURM_MEM_PER_NODE"]) / 1024
+            cpus = int(os.environ["SLURM_CPUS_PER_TASK"])
+        else:
+            ram_gb = psutil.virtual_memory().total / (1024**3)
 
-def memory_estimator(model, image_path, batch_size=16):
-    """estimate the inference latency and the memory usage
-    comenpared to the current slurm allocation"""
-    image_path = Path(image_path).expanduser()
-    if "SLURM_JOB_ID" in os.environ:
-        ram_gb = int(os.environ["SLURM_MEM_PER_NODE"]) / 1024
-        cpus = int(os.environ["SLURM_CPUS_PER_TASK"])
-    else:
-        ram_gb = psutil.virtual_memory().total / (1024**3)
-
-    inference_mb = inference_memory(model, batch_size)
-    latency_ms = model_latency_inference(model, image_path, batch_size)
-    print(
-        f"current memory allocation is {ram_gb} and the model inference estimate memory is {inference_mb} mb"
-    )
-    print(
-        f"the model inference latency for the provided data {image_path} is {latency_ms} ms"
-    )
-    return latency_ms, inference_mb
+        inference_mb = self.inference_memory(model, batch_size)
+        latency_ms = self.model_latency_inference(model, image_path, batch_size)
+        print(
+            f"current memory allocation is {ram_gb} and the model inference estimate memory is {inference_mb} mb"
+        )
+        print(
+            f"the model inference latency for the provided data {image_path} is {latency_ms} ms"
+        )
+        return latency_ms, inference_mb
